requirements.txt
Pillow==8.4.0
opencv-python
numpy
pandas
pytest
pytest-html
matplotlib

